anewlib: split anew into 2 files - anewlib and main.go files - to ensure package can be read by other files
